# backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: Shift Scheduler & Attendance API
  version: 2.4.0
  description: "Backend of docs/SPEC.md v2.4. Error envelope: { error: { code, message, details?, requestId } }. Lists: { data, meta: { page, limit, total } }. Calendar endpoints are range based (max 62 days)."
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
              example: EMPLOYEE_ON_TIME_OFF
            message:
              type: string
            details:
              type: array
              items:
                type: object
            requestId:
              type: string
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeNumber:
          type:
            - string
            - "null"
        firstName:
          type: string
        lastName:
          type: string
        email:
          type:
            - string
            - "null"
        phone:
          type:
            - string
            - "null"
        hourlyRate:
          type:
            - number
            - "null"
        payType:
          enum:
            - salary
            - hourly
          description: required on create
        publicHolidaysOff:
          type: boolean
          default: true
        status:
          enum:
            - active
            - on_leave
            - terminated
        employmentType:
          type: string
        birthDate:
          type:
            - string
            - "null"
        hiredOn:
          type:
            - string
            - "null"
        attendanceRequired:
          type: boolean
        workWeekdays:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 7
        terminatedOn:
          type:
            - string
            - "null"
        homeHotelId:
          type: integer
        hotels:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              isHome:
                type: boolean
        departments:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              color:
                type:
                  - string
                  - "null"
              hotelId:
                type: integer
...
paths:
...
  /employees:
    get:
      operationId: E1
      tags:
        - Employees
      summary: "E1: Employees assigned to a hotel (floating staff included)"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: hotelId
          in: query
          required: false
          schema:
            type: string
          description: Hotel id; required when the caller can access more than one hotel
        - name: departmentId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: string
          description: Page (default 1)
        - name: limit
          in: query
          required: false
          schema:
            type: string
          description: Page size (default 50, max 100)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: E3
      tags:
        - Employees
      summary: "E3: Create employee"
      description: "Role: admin."
      security:
        - bearer: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /employees/{id}:
    get:
      operationId: E2
      tags:
        - Employees
      summary: "E2: Employee (reduced view for non-home managers)"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: E4
      tags:
        - Employees
      summary: "E4: Update master data (home hotel managers)"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description: PRECONDITION_FAILED (If-Match)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: E5
      tags:
        - Employees
      summary: "E5: Soft-delete"
      description: "Role: admin."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      responses:
        "204":
          description: No content
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /employees/{id}/work-targets:
    get:
      operationId: E6
      tags:
        - Employees
      summary: "E6: Work targets"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: E7
      tags:
        - Employees
      summary: "E7: Set work targets"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /employees/{id}/hotels:
    put:
      operationId: E12
      tags:
        - Employees
      summary: "E12: Hotel assignments and home hotel"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /public-holidays:
    get:
      operationId: PH1
      tags:
        - Employees
      summary: "PH1: Public holidays of a hotel region"
      description: "Role: staff and up."
      security:
        - bearer: []
      parameters:
        - name: hotelId
          in: query
          required: false
          schema:
            type: string
          description: Hotel id; required when the caller can access more than one hotel
        - name: year
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...